#61 refactor: components/detail/comment 컴포넌트화 및 폴더구조 수정
- 답글 컴포넌트화 하였습니다.
- props setting도 하였습니다.

related:#61,#46
