+++ src/pages/[detail]/[title].tsx
@@ -5,7 +5,7 @@ import DetailHeader from '../../components/detail/detailHeader';
 import DetailArticle from '../../components/detail/detailArticle';
 import SectionUser from '../../components/detail/sectionUser';
 import SectionNevNext from '../../components/detail/sectionNevNext';
-import SectionComment from '../../components/detail/sectionComment';
+import SectionComment from '@/components/detail/sectioncomment/sectionComment';
 
 import detailStyle from '../../styles/detail/detail.module.css'
 import { useAppDispatch, useAppSelector } from '@/redux/hook';
